Added, Deleted, Updated and Changed Stuff

+++ src/commands/settings/antilinks.js
@@ -14,31 +14,49 @@ module.exports = {
 			);
 		}
 
+		const settings = await Guild.findOne({
+			guildID: message.guild.id,
+		});
+
 		if (args[0] === 'off') {
-			await Guild.updateOne(
-				{
-					guildID: message.guild.id,
-				},
-				{
-					'settings.antilinks': false,
-				},
-			);
-			message.channel.send(
-				'<:vSuccess:725270799098970112> Anti Links has been turned off',
-			).then(message.delete());
+			if(settings.settings.antilinks === false) {
+				return message.channel.send(
+					'<:vError:725270799124004934> Anti Links is already off.',
+				);
+			}
+			else {
+				await Guild.updateOne(
+					{
+						guildID: message.guild.id,
+					},
+					{
+						'settings.antilinks': false,
+					},
+				);
+				message.channel.send(
+					'<:vSuccess:725270799098970112> Anti Links has been turned off',
+				).then(message.delete());
+			}
 		}
 		else if (args[0] === 'on') {
-			await Guild.updateOne(
-				{
-					guildID: message.guild.id,
-				},
-				{
-					'settings.antilinks': true,
-				},
-			);
-			message.channel.send(
-				'<:vSuccess:725270799098970112> Anti Links has been turned on',
-			).then(message.delete());
+			if(settings.settings.antilinks === true) {
+				return message.channel.send(
+					'<:vError:725270799124004934> Anti Links is already on.',
+				);
+			}
+			else {
+				await Guild.updateOne(
+					{
+						guildID: message.guild.id,
+					},
+					{
+						'settings.antilinks': true,
+					},
+				);
+				message.channel.send(
+					'<:vSuccess:725270799098970112> Anti Links has been turned on',
+				).then(message.delete());
+			}
 		}
 	},
 };
